requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
google-genai==1.31.0
jinja2==3.1.6
python-multipart==0.0.20
requests-oauthlib==2.0.0
itsdangerous==2.2.0
